+++ server/seabattle-server/src/main/java/org/team/seabattle/server/service/validation/ImageShouldExists.java
