+++ server/seabattle-server/src/main/java/org/team/seabattle/server/view/RoomsController.java
@@ -1,15 +1,16 @@
 package org.team.seabattle.server.view;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import org.team.seabattle.server.view.dto.RegisterDto;
+import org.team.seabattle.server.view.dto.CreateRoomDto;
 
 @RestController
-public class AuthController {
+public class RoomsController {
 
-  public ResponseEntity<?> register(@RequestBody RegisterDto registerDto) {
-
+  @PostMapping("/api/v1/rooms")
+  public ResponseEntity<Long> createRoom(@RequestBody CreateRoomDto createRoomDto) {
     throw new UnsupportedOperationException();
   }
 }
