+++ server/seabattle-server/src/main/java/org/team/seabattle/server/view/dto/ImageDto.java
