+++ server/seabattle-server/src/main/resources/application.yaml
