+++ server/seabattle-server/src/test/resources/testimage1.jpg
