+++ server/seabattle-server/src/test/resources/testimage2.jpg
